Fadeout translucent drawable when collapsing
* Change default scrim color to none
* Fade out translucent when collapsingToolBarCollapsedColor is not defined
