+++ android/app/src/main/java/com/reactnativenavigation/params/parsers/CollapsingTopBarParamsParser.java
@@ -1,6 +1,5 @@
 package com.reactnativenavigation.params.parsers;
 
-import android.graphics.Color;
 import android.os.Bundle;
 
 import com.reactnativenavigation.params.CollapsingTopBarParams;
@@ -33,7 +32,7 @@ class CollapsingTopBarParamsParser extends Parser {
         result.imageUri = params.getString("collapsingToolBarImage", null);
         result.reactViewId = params.getString("collapsingToolBarComponent", null);
         result.expendOnTopTabChange = params.getBoolean("expendCollapsingToolBarOnTopTabChange");
-        result.scrimColor = getColor(params, "collapsingToolBarCollapsedColor", new StyleParams.Color(Color.WHITE));
+        result.scrimColor = getColor(params, "collapsingToolBarCollapsedColor", new StyleParams.Color());
         result.showTitleWhenCollapsed = hasReactView;
         result.collapseBehaviour = getCollapseBehaviour();
         return result;
